Allow customizing the CalciteDataSource via the QueryContextBuilder
