+++ core/api/src/main/java/com/blazebit/query/calcite/CalciteDataSource.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2024 - 2024 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.query.calcite;
+
+import java.sql.Driver;
+import javax.sql.DataSource;
+import org.apache.calcite.schema.SchemaPlus;
+
+/**
+ * @author Max Hovens
+ * @since 1.0.0
+ */
+public interface CalciteDataSource extends Driver, DataSource {
+
+    /**
+     * Returns the root schema of the Calcite JDBC DataSource.
+     *
+     * @return The root schema of the Calcite JDBC DataSource
+     */
+    SchemaPlus getRootSchema();
+}
