+++ core/api/src/main/java/com/blazebit/query/calcite/CalciteDataSourceBuilder.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2024 - 2024 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.query.calcite;
+
+/**
+ * @author Max Hovens
+ * @since 1.0.0
+ */
+public interface CalciteDataSourceBuilder {
+
+    /**
+     * Sets a property for the {@link CalciteDataSource}.
+     *
+     * @param name The name of the property to be set
+     * @param value The value of the property to be set
+     * @return The CalciteDataSourceBuilder instance
+     */
+    CalciteDataSourceBuilder setProperty(String name, String value);
+
+    /**
+     * Builds a {@link CalciteDataSource}.
+     *
+     * @return The {@link CalciteDataSource} instance built.
+     */
+    CalciteDataSource build();
+}
